make input handlers and transform updates cleaner
