+++ src/main/java/me/lsdo/processing/PixelMeshAnimation.java
@@ -19,7 +19,6 @@ public abstract class PixelMeshAnimation<T extends LedPixel> {
     public double frameRate = 0.;  // fps
 
     InputControl ctrl;
-    boolean txChanged = false;
     
     public PixelMeshAnimation(PixelMesh<? extends T> mesh) {
         this.mesh = mesh;
@@ -29,291 +28,11 @@ public abstract class PixelMeshAnimation<T extends LedPixel> {
     void initControl() {
         ctrl = new InputControl();
 	ctrl.init();
-
-        ctrl.registerHandler("jog_a", new InputControl.InputHandler() {
-		@Override
-                public void jog(boolean pressed) {
-                    boolean forward = pressed;
-		    mesh.placement.xo += (forward ? 1 : -1) * .01;
-		    txChanged = true;
-                }
-            });
-        ctrl.registerHandler("jog-xo", new InputControl.InputHandler() {
-		@Override
-                public void jog(boolean pressed) {
-                    boolean forward = pressed;
-		    mesh.placement.xo += (forward ? 1 : -1) * .01;
-		    txChanged = true;
-                }
-            });
-        ctrl.registerHandler("xo", new InputControl.InputHandler() {
-		@Override
-                public void set(double d) {
-		    mesh.placement.xo = d;
-		    txChanged = true;
-                }
-            });
-        ctrl.registerHandler("jog_b", new InputControl.InputHandler() {
-		@Override
-                public void jog(boolean pressed) {
-                    boolean forward = pressed;
-		    mesh.placement.yo += (forward ? 1 : -1) * .01;
-		    txChanged = true;
-                }
-            });
-        ctrl.registerHandler("jog-yo", new InputControl.InputHandler() {
-		@Override
-                public void jog(boolean pressed) {
-                    boolean forward = pressed;
-		    mesh.placement.yo += (forward ? 1 : -1) * .01;
-		    txChanged = true;
-                }
-            });
-        ctrl.registerHandler("yo", new InputControl.InputHandler() {
-		@Override
-                public void set(double d) {
-		    mesh.placement.yo = d;
-		    txChanged = true;
-                }
-            });
-        ctrl.registerHandler("pitch_a", new InputControl.InputHandler() {
-		@Override
-                public void slider(double val) {
-		    mesh.placement.rot = 2*Math.PI*(val - .5);
-		    txChanged = true;
-                }
-            });
-        ctrl.registerHandler("jog-rot", new InputControl.InputHandler() {
-		@Override
-                public void jog(boolean pressed) {
-                    boolean forward = pressed;
-		    mesh.placement.rot += (forward ? 1 : -1) * .01 * Math.PI;
-		    txChanged = true;
-                }
-            });
-        ctrl.registerHandler("rot", new InputControl.InputHandler() {
-		@Override
-                public void set(double d) {
-		    mesh.placement.rot = Math.toRadians(d);
-		    txChanged = true;
-                }
-            });
-        ctrl.registerHandler("pitch_b", new InputControl.InputHandler() {
-		@Override
-                public void slider(double val) {
-		    mesh.placement.scale = Math.exp(2*(val - .5));
-		    txChanged = true;
-                }
-            });
-        ctrl.registerHandler("jog-scale", new InputControl.InputHandler() {
-		@Override
-                public void jog(boolean pressed) {
-                    boolean forward = pressed;
-		    mesh.placement.scale *= 1. + (forward ? 1 : -1) * .01;
-		    txChanged = true;
-                }
-            });
-        ctrl.registerHandler("scale", new InputControl.InputHandler() {
-		@Override
-                public void set(double d) {
-		    mesh.placement.scale = d;
-		    txChanged = true;
-                }
-            });
-        ctrl.registerHandler("playpause_a", new InputControl.InputHandler() {
-		@Override
-                public void button(boolean pressed) {
-                    if (pressed) {
-			if (mesh instanceof Prometheus) {
-			    Prometheus p = (Prometheus)mesh;
-			    Prometheus.WingDisplayMode[] modes = Prometheus.WingDisplayMode.values();
-			    for (int i = 0; i < modes.length; i++) {
-				if (p.mode == modes[i]) {
-				    p.mode = modes[(i + 1) % modes.length];
-				    System.out.println("setting wing mode " + p.mode.name());
-				    break;
-				}
-			    }
-			    txChanged = true;
-			}
-                    }
-                }
-            });
-        ctrl.registerHandler("wingmode", new InputControl.InputHandler() {
-		@Override
-                public void set(String s) {
-		    if (mesh instanceof Prometheus) {
-			Prometheus p = (Prometheus)mesh;
-			if (s.equals("unified")) {
-			    p.mode = Prometheus.WingDisplayMode.UNIFIED;
-			} else if (s.equals("mirror")) {
-			    p.mode = Prometheus.WingDisplayMode.MIRROR;
-			} else if (s.equals("flip")) {
-			    p.mode = Prometheus.WingDisplayMode.FLIP_HORIZ;
-			} else if (s.equals("opposite")) {
-			    p.mode = Prometheus.WingDisplayMode.ROTATE_180;
-			} else {
-			    return;
-			}
-			txChanged = true;
-                    }
-                }
-            });
-        ctrl.registerHandler("playpause_b", new InputControl.InputHandler() {
-		@Override
-                public void button(boolean pressed) {
-		    if (!(mesh instanceof Prometheus)) {
-			return;
-		    }
-		    Prometheus prom = ((Prometheus)mesh);
-		    if (pressed) {
-			prom.startFlapping();
-		    } else {
-			prom.stopFlapping();
-		    }
-		    txChanged = true;
-                }
-            });
-        ctrl.registerHandler("flap", new InputControl.InputHandler() {
-		@Override
-                public void set(boolean pressed) {
-		    if (!(mesh instanceof Prometheus)) {
-			return;
-		    }
-		    Prometheus prom = ((Prometheus)mesh);
-		    if (pressed) {
-			prom.startFlapping();
-		    } else {
-			prom.stopFlapping();
-		    }
-		    txChanged = true;
-                }
-            });
-        ctrl.registerHandler("mixer", new InputControl.InputHandler() {
-		@Override
-                public void slider(double val) {
-		    if (!(mesh instanceof Prometheus)) {
-			return;
-		    }
-		    Prometheus prom = ((Prometheus)mesh);
-		    prom.setFlapAngle(prom.minFlapAngle * (1 - val) + prom.maxFlapAngle * val);
-		    txChanged = true;
-                }
-            });
-        ctrl.registerHandler("flap-angle", new InputControl.InputHandler() {
-		@Override
-                public void slider(double val) {
-		    if (!(mesh instanceof Prometheus)) {
-			return;
-		    }
-		    Prometheus prom = ((Prometheus)mesh);
-		    prom.setFlapAngle(prom.minFlapAngle * (1 - val) + prom.maxFlapAngle * val);
-		    txChanged = true;
-                }
-            });
-        ctrl.registerHandler("flap-depth", new InputControl.InputHandler() {
-		@Override
-                public void slider(double val) {
-		    if (!(mesh instanceof Prometheus)) {
-			return;
-		    }
-		    Prometheus prom = ((Prometheus)mesh);
-		    prom.maxFlap = prom.minMaxFlap * (1 - val) + prom.maxMaxFlap * val;
-		    txChanged = true;
-                }
-            });
-        ctrl.registerHandler("flap-speed", new InputControl.InputHandler() {
-		@Override
-                public void slider(double val) {
-		    if (!(mesh instanceof Prometheus)) {
-			return;
-		    }
-		    Prometheus prom = ((Prometheus)mesh);
-		    prom.flapPeriod = prom.maxFlapPeriod * Math.pow(prom.minFlapPeriod / prom.maxFlapPeriod, val);
-		    txChanged = true;
-                }
-            });
-        ctrl.registerHandler("load_a", new InputControl.InputHandler() {
-		@Override
-                public void button(boolean pressed) {
-		    if (pressed && PixelMeshAnimation.this instanceof WindowAnimation) {
-			WindowAnimation win = (WindowAnimation)PixelMeshAnimation.this;
-			win.preserveAspect = !win.preserveAspect;
-			txChanged = true;
-		    }
-                }
-            });
-        ctrl.registerHandler("stretch", new InputControl.InputHandler() {
-		@Override
-                public void set(boolean b) {
-		    if (PixelMeshAnimation.this instanceof WindowAnimation) {
-			WindowAnimation win = (WindowAnimation)PixelMeshAnimation.this;
-			win.preserveAspect = !b;
-			txChanged = true;
-		    }
-                }
-            });
-        ctrl.registerHandler("wingmode_unified", new InputControl.InputHandler() {
-		@Override
-                public void set(boolean pressed) {
-		    if (pressed) {
-			((Prometheus)mesh).mode = Prometheus.WingDisplayMode.UNIFIED;
-		    }
-		    txChanged = true;
-		}
-            });
-        ctrl.registerHandler("wingmode_mirror", new InputControl.InputHandler() {
-		@Override
-                public void set(boolean pressed) {
-		    if (pressed) {
-			((Prometheus)mesh).mode = Prometheus.WingDisplayMode.MIRROR;
-		    }
-		    txChanged = true;
-		}
-            });
-        ctrl.registerHandler("wingmode_flip", new InputControl.InputHandler() {
-		@Override
-                public void set(boolean pressed) {
-		    if (pressed) {
-			((Prometheus)mesh).mode = Prometheus.WingDisplayMode.FLIP_HORIZ;
-		    }
-		    txChanged = true;
-		}
-            });
-        ctrl.registerHandler("wingmode_rotate", new InputControl.InputHandler() {
-		@Override
-                public void set(boolean pressed) {
-		    if (pressed) {
-			((Prometheus)mesh).mode = Prometheus.WingDisplayMode.ROTATE_180;
-		    }
-		    txChanged = true;
-		}
-            });
-        ctrl.registerHandler("stretch_yes", new InputControl.InputHandler() {
-		@Override
-                public void set(boolean pressed) {
-		    if (pressed) {
-			if (PixelMeshAnimation.this instanceof WindowAnimation) {
-			    WindowAnimation win = (WindowAnimation)PixelMeshAnimation.this;
-			    win.preserveAspect = false;
-			    txChanged = true;
-			}
-		    }
-		}
-            });
-        ctrl.registerHandler("stretch_no", new InputControl.InputHandler() {
-		@Override
-                public void set(boolean pressed) {
-		    if (pressed) {
-			if (PixelMeshAnimation.this instanceof WindowAnimation) {
-			    WindowAnimation win = (WindowAnimation)PixelMeshAnimation.this;
-			    win.preserveAspect = true;
-			    txChanged = true;
-			}
-		    }
-		}
-            });
+	mesh.registerHandlers(ctrl);
+	this.registerHandlers(ctrl);
     }
+
+    public void registerHandlers(InputControl ctrl) {}
     
     public void draw(double t) {
 	if (!initialized) {
@@ -321,27 +40,12 @@ public abstract class PixelMeshAnimation<T extends LedPixel> {
 	    initialized = true;
 	}
 
-	ctrl.processInput();
-	// hacky
-	if (mesh instanceof Prometheus && ((Prometheus)mesh).manageFlapState(this)) {
-	    txChanged = true;
-	}
-	if (txChanged) {
-	    if (this instanceof PixelTransform.TransformListener) {
-		((PixelTransform.TransformListener)this).transformChanged();
-	    }
-	    txChanged = false;
-	    /*
-	    System.out.println("xo: " + mesh.placement.xo);
-	    System.out.println("yo: " + mesh.placement.yo);
-	    System.out.println("rot: " + mesh.placement.rot);
-	    System.out.println("scale: " + mesh.placement.scale);
-	    */
-	}
-	
 	double deltaT = t - lastT;
 	lastT = t;
 	updateFramerate(deltaT);
+		
+	ctrl.processInput();
+	mesh.beforeDraw(this);
 	
         preFrame(t, deltaT);
         for (T c : mesh.coords){
